refactor(tests): extract Google adapter tests into test_runner_google.py
test_runner_api_loop.py exceeded the 1000-line convention limit (was 1113).
Move TestRunGoogle, TestGoogleAdapter, and TestTranslateMessagesGoogle into
a dedicated tests/test_runner_google.py — keeping api loop tests focused on
loop lifecycle and loop manager behaviour.
