+++ tests/test_runner_api_loop.py
@@ -8,7 +8,6 @@ from unittest.mock import MagicMock, patch
 
 from theforge.agent_types import ModelUsage
 from theforge.config import ModelProfile
-from theforge.runners.adapters.google import _make_google_adapter
 from theforge.runners.api import (
     AgentLoopManager,
     run_api_agent,
@@ -693,354 +692,6 @@ class TestAgentLoopLifecycle:
         assert not result.success
         assert "Provider API error" in result.output
         assert "'NoneType' object is not iterable" in result.output
-
-    def _make_google_modules(self, fake_client):
-        """Build sys.modules mocks for google.genai, wiring attributes correctly.
-
-        `import google.genai as genai` resolves via attribute access on the google
-        module object, not directly from sys.modules["google.genai"], so we must set
-        mock_google.genai = mock_genai and mock_genai.types = mock_genai_types.
-        """
-        import sys
-
-        mock_genai_types = MagicMock()
-        mock_genai = MagicMock()
-        mock_genai.Client.return_value = fake_client
-        mock_genai.types = mock_genai_types
-
-        mock_google = MagicMock()
-        mock_google.genai = mock_genai
-
-        modules = {
-            "google": mock_google,
-            "google.genai": mock_genai,
-            "google.genai.types": mock_genai_types,
-        }
-        return sys.modules, modules, mock_genai, mock_genai_types
-
-    def test_google_adapter_fc_args_none_parsed_as_empty_dict(self, tmp_path):
-        """Google adapter with fc.args = None should produce arguments={}."""
-        # Build a fake response where a function_call part has args=None
-        fake_fc = MagicMock()
-        fake_fc.name = "submit_review"
-        fake_fc.args = None
-
-        fake_part = MagicMock()
-        fake_part.function_call = fake_fc
-        fake_part.text = None
-
-        fake_candidate = MagicMock()
-        fake_candidate.content = MagicMock()
-        fake_candidate.content.parts = [fake_part]
-
-        fake_response = MagicMock()
-        fake_response.candidates = [fake_candidate]
-        fake_response.usage_metadata = None
-        fake_response.prompt_feedback = None
-
-        fake_client = MagicMock()
-        fake_client.models.generate_content.return_value = fake_response
-
-        import sys
-
-        _, modules, _, _ = self._make_google_modules(fake_client)
-        profile = _make_profile(provider="google", model="gemini-2.5-flash")
-
-        with patch.dict(sys.modules, modules):
-            adapter = _make_google_adapter(profile, secrets=None)
-            turn = adapter([{"role": "user", "content": "hi"}], [])
-
-        assert len(turn.tool_calls) == 1
-        assert turn.tool_calls[0].arguments == {}
-
-    def test_google_adapter_content_parts_none_yields_empty(self, tmp_path):
-        """Google adapter with candidate.content.parts = None → no tool_calls, no text."""
-        fake_candidate = MagicMock()
-        fake_candidate.content = MagicMock()
-        fake_candidate.content.parts = None  # parts is None, not []
-
-        fake_response = MagicMock()
-        fake_response.candidates = [fake_candidate]
-        fake_response.usage_metadata = None
-        fake_response.prompt_feedback = None
-
-        fake_client = MagicMock()
-        fake_client.models.generate_content.return_value = fake_response
-
-        import sys
-
-        _, modules, _, _ = self._make_google_modules(fake_client)
-        profile = _make_profile(provider="google", model="gemini-2.5-flash")
-
-        with patch.dict(sys.modules, modules):
-            adapter = _make_google_adapter(profile, secrets=None)
-            turn = adapter([{"role": "user", "content": "hi"}], [])
-
-        assert turn.tool_calls == []
-        assert turn.text_output is None
-
-    def test_google_adapter_thought_signature_attached_to_fc(self, tmp_path):
-        """Function calls are tagged with thought_signature from preceding thought part."""
-        import sys
-
-        fake_thought_part = MagicMock()
-        fake_thought_part.thought = True
-        fake_thought_part.thought_signature = "sig-abc"
-        fake_thought_part.function_call = None
-        fake_thought_part.text = None
-
-        fake_fc = MagicMock()
-        fake_fc.name = "submit_review"
-        fake_fc.args = {"verdict": "APPROVE"}
-        fake_fc.thought_signature = None  # not on the fc itself
-
-        fake_fc_part = MagicMock()
-        fake_fc_part.thought = False
-        fake_fc_part.function_call = fake_fc
-        fake_fc_part.text = None
-
-        fake_candidate = MagicMock()
-        fake_candidate.content = MagicMock()
-        fake_candidate.content.parts = [fake_thought_part, fake_fc_part]
-        fake_candidate.finish_reason = "STOP"
-
-        fake_response = MagicMock()
-        fake_response.candidates = [fake_candidate]
-        fake_response.usage_metadata = None
-        fake_response.prompt_feedback = None
-
-        fake_client = MagicMock()
-        fake_client.models.generate_content.return_value = fake_response
-
-        _, modules, _, _ = self._make_google_modules(fake_client)
-        profile = _make_profile(provider="google", model="gemini-2.5-pro-preview-customtools")
-
-        with patch.dict(sys.modules, modules):
-            adapter = _make_google_adapter(profile, secrets=None)
-            turn = adapter([{"role": "user", "content": "review please"}], [])
-
-        assert len(turn.tool_calls) == 1
-        assert turn.tool_calls[0].thought_signature == "sig-abc"
-
-    def test_google_translate_messages_includes_thought_signature(self, tmp_path):
-        """_translate_messages_google embeds thought_signature in function_call parts."""
-        from theforge.runners.adapters.google import _translate_messages_google
-
-        tc = ToolCallRequest(
-            id="call_0",
-            name="submit_review",
-            arguments={"verdict": "APPROVE"},
-            thought_signature="sig-xyz",
-        )
-        messages = [{"role": "assistant", "tool_calls": [tc], "content": None}]
-        result = _translate_messages_google(messages)
-
-        assert len(result) == 1
-        fc_part = result[0]["parts"][0]["function_call"]
-        assert fc_part["thought_signature"] == "sig-xyz"
-
-    def test_google_translate_messages_no_thought_signature_when_absent(self, tmp_path):
-        """_translate_messages_google omits thought_signature when None."""
-        from theforge.runners.adapters.google import _translate_messages_google
-
-        tc = ToolCallRequest(
-            id="call_0",
-            name="submit_review",
-            arguments={"verdict": "APPROVE"},
-            thought_signature=None,
-        )
-        messages = [{"role": "assistant", "tool_calls": [tc], "content": None}]
-        result = _translate_messages_google(messages)
-
-        fc_part = result[0]["parts"][0]["function_call"]
-        assert "thought_signature" not in fc_part
-
-
-class TestRunGoogle:
-    """Tests for the _run_google single-shot API path."""
-
-    def _make_google_modules(self, fake_client):
-        """Build sys.modules mocks for google.genai."""
-        import sys
-
-        mock_genai_types = MagicMock()
-        mock_genai = MagicMock()
-        mock_genai.Client.return_value = fake_client
-        mock_genai.types = mock_genai_types
-
-        mock_google = MagicMock()
-        mock_google.genai = mock_genai
-
-        modules = {
-            "google": mock_google,
-            "google.genai": mock_genai,
-            "google.genai.types": mock_genai_types,
-        }
-        return sys.modules, modules
-
-    def _make_response(
-        self,
-        text=None,
-        block_reason=None,
-        finish_reason="STOP",
-        input_tokens=100,
-        output_tokens=50,
-    ):
-        """Build a minimal fake Gemini response object."""
-        usage = MagicMock()
-        usage.prompt_token_count = input_tokens
-        usage.candidates_token_count = output_tokens
-
-        candidate = MagicMock()
-        candidate.finish_reason = finish_reason
-
-        feedback = MagicMock()
-        feedback.block_reason = block_reason
-
-        response = MagicMock()
-        response.text = text
-        response.usage_metadata = usage
-        response.candidates = [candidate]
-        response.prompt_feedback = feedback
-        return response
-
-    def test_valid_json_response_succeeds(self):
-        """Normal JSON response returns success with structured_data parsed."""
-        import sys
-
-        from theforge.runners.adapters.google import _run_google
-
-        valid_json = (
-            '{"verdict": "APPROVE", "summary": "ok", "findings": [],'
-            ' "story_compliance": {"matches_spec": true, "mismatches": []},'
-            ' "test_coverage": {"adequate": true, "gaps": []}}'
-        )
-        fake_response = self._make_response(text=valid_json)
-        fake_client = MagicMock()
-        fake_client.models.generate_content.return_value = fake_response
-
-        _, modules = self._make_google_modules(fake_client)
-        profile = _make_profile(provider="google", model="gemini-2.5-flash")
-
-        with patch.dict(sys.modules, modules):
-            result = _run_google("test prompt", profile)
-
-        assert result.success
-        assert result.exit_code == 0
-        assert result.output == valid_json
-        assert result.structured_data is not None
-
-    def test_none_response_text_returns_clear_error(self):
-        """response.text is None → clear error, not TypeError."""
-        import sys
-
-        from theforge.runners.adapters.google import _run_google
-
-        # Both calls return None (initial + retry)
-        fake_response = self._make_response(text=None, finish_reason="STOP", input_tokens=1234)
-        fake_client = MagicMock()
-        fake_client.models.generate_content.return_value = fake_response
-
-        _, modules = self._make_google_modules(fake_client)
-        profile = _make_profile(provider="google", model="gemini-2.5-flash")
-
-        with patch.dict(sys.modules, modules):
-            result = _run_google("test prompt", profile)
-
-        assert not result.success
-        assert result.exit_code == 1
-        assert "TypeError" not in result.output
-        assert "empty" in result.output.lower() or "retry" in result.output.lower()
-
-    def test_blocked_response_surfaces_block_reason(self):
-        """Block reason is included in the error message."""
-        import sys
-
-        from theforge.runners.adapters.google import _run_google
-
-        fake_response = self._make_response(text=None, block_reason="SAFETY", input_tokens=500)
-        fake_client = MagicMock()
-        fake_client.models.generate_content.return_value = fake_response
-
-        _, modules = self._make_google_modules(fake_client)
-        profile = _make_profile(provider="google", model="gemini-2.5-flash")
-
-        with patch.dict(sys.modules, modules):
-            result = _run_google("test prompt", profile)
-
-        assert not result.success
-        assert "SAFETY" in result.output
-        assert "block_reason" in result.output
-        assert "500" in result.output  # input token count
-
-    def test_non_stop_finish_reason_in_error(self):
-        """Non-STOP finish_reason (SAFETY, RECITATION, MAX_TOKENS) is logged explicitly."""
-        import sys
-
-        from theforge.runners.adapters.google import _run_google
-
-        for finish_reason in ("SAFETY", "RECITATION", "MAX_TOKENS"):
-            fake_response = self._make_response(
-                text=None, block_reason=None, finish_reason=finish_reason, input_tokens=200
-            )
-            fake_client = MagicMock()
-            fake_client.models.generate_content.return_value = fake_response
-
-            _, modules = self._make_google_modules(fake_client)
-            profile = _make_profile(provider="google", model="gemini-2.5-flash")
-
-            with patch.dict(sys.modules, modules):
-                result = _run_google("test prompt", profile)
-
-            assert not result.success, f"Expected failure for finish_reason={finish_reason}"
-            assert finish_reason in result.output, f"Expected {finish_reason} in output"
-            assert "200" in result.output  # input token count
-
-    def test_empty_unblocked_response_retries_once(self):
-        """Empty unblocked response (STOP) triggers one retry before giving up."""
-        import sys
-
-        from theforge.runners.adapters.google import _run_google
-
-        empty_response = self._make_response(text=None, finish_reason="STOP")
-        fake_client = MagicMock()
-        fake_client.models.generate_content.return_value = empty_response
-
-        _, modules = self._make_google_modules(fake_client)
-        profile = _make_profile(provider="google", model="gemini-2.5-flash")
-
-        with patch.dict(sys.modules, modules):
-            result = _run_google("test prompt", profile)
-
-        # Should have called generate_content twice (initial + retry)
-        assert fake_client.models.generate_content.call_count == 2
-        assert not result.success
-
-    def test_retry_succeeds_on_second_attempt(self):
-        """If retry returns valid text, the result is successful."""
-        import sys
-
-        from theforge.runners.adapters.google import _run_google
-
-        valid_json = (
-            '{"verdict": "APPROVE", "summary": "ok", "findings": [],'
-            ' "story_compliance": {"matches_spec": true, "mismatches": []},'
-            ' "test_coverage": {"adequate": true, "gaps": []}}'
-        )
-        empty_response = self._make_response(text=None, finish_reason="STOP")
-        good_response = self._make_response(text=valid_json)
-        fake_client = MagicMock()
-        fake_client.models.generate_content.side_effect = [empty_response, good_response]
-
-        _, modules = self._make_google_modules(fake_client)
-        profile = _make_profile(provider="google", model="gemini-2.5-flash")
-
-        with patch.dict(sys.modules, modules):
-            result = _run_google("test prompt", profile)
-
-        assert fake_client.models.generate_content.call_count == 2
-        assert result.success
-        assert result.structured_data is not None
 
 
 class TestTimeNudge:
